rank learning factor stat, fixes, changes, prettyfi

+++ 1.6/Source/Core40k/Rank/RankData.cs
@@ -7,7 +7,7 @@ public class RankData
 {
     public RankDef rankDef;
 
-    public int daysAs = 0;
+    public float daysAs = 0f;
 
     public RankData()
     {
